% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InCHlib.R
\name{renderInCHlib}
\alias{renderInCHlib}
\title{Widget render function for use in Shiny}
\usage{
renderInCHlib(expr, env = parent.frame(), quoted = FALSE)
}
\description{
Widget render function for use in Shiny
}

